requirements.txt
openai
python-dotenv
rich
black
ruff
pyaudio
wave
websockets
openai-whisper
webrtcvad # For pause detection